// client/src/app/models/hackerNews.model.ts
export interface ItemDto
    {
        id: number;
        deleted: boolean;
        type?: string;
        by?: string;
        time: number;
        text?: string;
        dead: boolean;
        parent?: number;
        poll?: number;
        kids?: number[];
        url?: string;
        score?: number;
        title?: string;
        parts?: number[];
        descendants?: number;
    }

export interface PagedResult<T>{
    items: T[];
    page: number;
    pageSize: number;
    totalCount: number;
    hasNext: boolean;
}